refactor:auth msw file moved
